Port existing heuristics to TD framework
[ghstack-poisoned]
